[DATALAD] Recorded changes
